avoid XPath access when linear search makes more sense
The XPath parsing cannot be done only once, and is fairly expensive
in REXML (maybe a case of changing XML parser again :().

In any case, for elements that have a small number of children, use
linear search, which turns out to be significantly faster.

One functionality change is that #each_world does not accept
a recursive flag anymore - as if there is a world it must be
toplevel - and each_model(recursive: true) will only search
within a world.