match behaviour of model-included-in-model with sdformat
Gazebo actually merges the content of the included model in the
enclosing one, instead of adding the included model in it.

In effect

  <model name="test">
     <include>
        <uri>TEST_URI</uri>
     </include>
  </model>

is equivalent to

  <include>
    <name>test</name>
    <uri>TEST_URI</uri>
  </include>

To achieve this, we must transform the included tree to put the Model
elements in the parent model, transforming their names and poses
according to the include tag's pose and name attribute.