add support to handle "not flattened" model-in-model structures
So far, models in models were "flattened", that is that their elements
were added namespaced and added to the parent model.

This adds support to avoid this flattening. While flattening is great
for things that only care about kinematics, it is problematic when
trying to use the semantics (i.e. finding the arms in a robot that
has a combination of two arms)