unit-test and fix flatten_model_tree
The joints pose is specified w.r.t. the parent (i.e. the model), and
must therefore be transformed. However, the joint axis is normally
specified w.r.t. the joint frame. In any case, it specifies a direction
and must therefore only be rotated if use_model_parent_frame is set.