fix corner-cases in parsing complex elements
This commit improves error reporting when complex elements are not
valid. It adds explicit checks on the number of elements (i.e. giving
a proper error message when a 3-vector has only two elements, instead
of falling into a TypeError later on).

Moreover, it fixes vector3_to_eigen to strip the text before parsing.